# src/infra/rest-api/endpoints/openapi.yaml
openapi: 3.1.0
info:
  title: sump api
  description: sump - simple user management platform
  version: 0.1.0
...
components:
  schemas:
...
    CreateNewEnvironmentOperationDto:
      type: object
      required:
        - name
        - customProperties
      properties:
        name:
          type: string
        customProperties:
          type: object
...
    SetEnvironmentCustomPropertyOperationDto:
      type: object
    DeleteEnvironmentCustomPropertyOperationDto:
      type: object
...
    UpdateEnvironmentByIdAndTenantIdOperationDto:
      type: object
      properties:
        name:
          type: string
...
    CreateNewEnvironmentOperationDtoResponse:
      type: object
      properties:
        environmentId:
          type: string
...
    GetEnvironmentByIdAndTenantIdOperationDtoResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        tenantId:
          type: string
        customProperties:
          type: object
...
paths:
...
  /api/v1/tenants/{tenantId}/environments:
    post:
      operationId: createnewenvironment
      tags:
        - Tenant Environment
      parameters:
        - name: tenantId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNewEnvironmentOperationDto'
      responses:
        '201':
          description: 'Expected response'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateNewEnvironmentOperationDtoResponse'
  /api/v1/tenants/{tenantId}/environments/{environmentId}:
    patch:
      operationId: updateenvironmentbyidandtenantid
      tags:
        - Tenant Environment
      parameters:
        - name: tenantId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
        - name: environmentId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEnvironmentByIdAndTenantIdOperationDto'
      responses:
        '200':
          description: 'Expected response'
    delete:
      operationId: deleteenvironmentbyidandtenantid
      tags:
        - Tenant Environment
      parameters:
        - name: tenantId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
        - name: environmentId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Expected response'
    get:
      operationId: getenvironmentbyidandtenantid
      tags:
        - Tenant Environment
      parameters:
        - name: tenantId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
        - name: environmentId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Expected response'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetEnvironmentByIdAndTenantIdOperationDtoResponse'
  /api/v1/tenants/{tenantId}/environments/{environmentId}/custom-property:
    patch:
      operationId: updateenvironmentcustompropertiesbyidandtenantid
      tags:
        - Tenant Environment
      parameters:
        - name: tenantId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
        - name: environmentId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetEnvironmentCustomPropertyOperationDto'
      responses:
        '200':
          description: 'Expected response'

    delete:
      operationId: deleteenvironmentcustompropertiesbyidandtenantid
      tags:
        - Tenant Environment
      parameters:
        - name: tenantId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
        - name: environmentId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteEnvironmentCustomPropertyOperationDto'
      responses:
        '200':
          description: 'Expected response'
...